ParkingLot: Implemented unpark functionality and HourlyFeeStrategy

+++ src/Problems/parkinglot/solution/strategy/fee/FeeStrategy.java
@@ -3,5 +3,5 @@ package Problems.parkinglot.solution.strategy.fee;
 import Problems.parkinglot.solution.model.ParkingTicket;
 
 public interface FeeStrategy {
-    Double calculateFee(ParkingTicket ticket);
+    Float calculateFee(ParkingTicket ticket);
 }
